requirements.txt
streamlit
geopandas
pandas
shapely
folium
streamlit-folium
plotly-express
numpy
openpyxl
python-docx
branca
matplotlib
shap
